refactor(TicketMachine, TicketManager) : 역할 구분 모호하여 구조 재구성을 위해 클래스 결합
