+++ src/main/java/edu/nextstep/lottoauto/domain/TicketMachine.java
@@ -1,6 +1,7 @@
 package edu.nextstep.lottoauto.domain;
 
 import edu.nextstep.lottoauto.domain.ticketmaker.CustomNumbersMaker;
+import edu.nextstep.lottoauto.exception.PaymentIllegalArgumentException;
 import edu.nextstep.lottoauto.view.form.WinningResultForm;
 import edu.nextstep.lottoauto.domain.ticketmaker.NumbersMaker;
 
@@ -8,16 +9,25 @@ import java.util.*;
 
 public class TicketMachine {
 
-    private static final int A_UNIT_PRICE = 1_000;
+    private static final int TICKET_PRICE = 1_000;
 
     private final TicketRepository ticketRepository = new TicketRepository();
 
-    public void createAndSaveTickets(int numberOfTickets, NumbersMaker numbersMaker) {
+    public void createAndSaveTickets(int payment, NumbersMaker numbersMaker) {
+        validateUnderAUnitPrice(payment);
+        validateDivideUnitPrice(payment);
+
+        int numberOfTickets = calculateNumberOfTicketsFrom(payment);
         for (int i = 0; i < numberOfTickets; i++) {
             Ticket ticket = Ticket.madeBy(numbersMaker);
             ticketRepository.save(ticket);
         }
     }
+
+    public int calculateNumberOfTicketsFrom(int payment) {
+        return payment / TICKET_PRICE;
+    }
+
 
     public List<Ticket> findTickets() {
         return ticketRepository.findAll();
@@ -54,10 +64,27 @@ public class TicketMachine {
     }
 
     private int calculatePayment(List<Ticket> tickets) {
-        return tickets.size() * A_UNIT_PRICE;
+        return tickets.size() * TICKET_PRICE;
+    }
+
+    public WinningResultForm confirmWinningResult(String winningNumbersString) {
+        Ticket winningTicket = createCustomTicket(winningNumbersString);
+        return confirmWinningResult(winningTicket);
     }
 
     public Ticket createCustomTicket(String winningNumbersString) {
         return Ticket.madeBy(new CustomNumbersMaker(winningNumbersString));
+    }
+
+    private void validateUnderAUnitPrice(int payment) {
+        if (payment < TICKET_PRICE) {
+            throw new PaymentIllegalArgumentException("최소 입력 가능 금액 미달. 최소 입력 금액 : " + TICKET_PRICE + " 원");
+        }
+    }
+
+    private void validateDivideUnitPrice(int payment) {
+        if ((payment % TICKET_PRICE) != 0) {
+            throw new PaymentIllegalArgumentException("개 당 금액 " + TICKET_PRICE + " 원으로 해당 단위로 입력 필요.");
+        }
     }
 }
